ai-expense-reporter/src/upload/requirements.txt
json
boto3
python-dotenv